+++ things/redis/streams.go
@@ -7,7 +7,7 @@ import (
 	"context"
 
 	"github.com/go-redis/redis"
-	"github.com/mainflux/mainflux/pkg/groups"
+	"github.com/mainflux/mainflux/internal/groups"
 	"github.com/mainflux/mainflux/things"
 )
